(#69) Fix query table name escaping
Previously, table names were escaped by doing a global search and
replace of the table name throughout the entire query.  So if the table
name appeared _anywhere_ within the query, it would get escaped.  This
was most obvious when selecting from `-` (_i.e._ stdin).  For instance,
a query like this:
```sql
select 'foo-bar', * from -
```
would be parsed, and `-` identified as a table name.  All occurrences of
`-` in the query would be replaced with `` `-` ``, resulting in the query
```sql
select 'foo`-`bar', * from `-`
```
The results would look like this:
```
+-------------+-------+---------+
| 'foo`-`bar' |  c1   |   c2    |
+-------------+-------+---------+
| foo`-`bar   | other | columns |
+-------------+-------+---------+
```

This update ensures that the substitutions are done only to the
referenced tables in the query.
1. Update `sqlFields()` to parse _entire_ query into an array.
1. Update `TableNames()` to walk through the parsed query, identify
   tables, and make a note of their position within the parsed query
   array.
1. Update `Import()` to make replacements to the identified elements of
   the parsed query, and then re-assemble the query by joining the array
   back into a string.